Blocking message sender open; added AZURE_NODISCARD to help ensure APIs that are likely to fail have their returns checked
