+++ sdk/core/azure-core-amqp/samples/internal/eventhub_get_properties_sample/eventhub_get_eventhub_properties_sample.cpp
@@ -142,7 +142,7 @@ std::tuple<bool, EventHubPartitionProperties> GetPartitionProperties(
       throw std::runtime_error("Unexpected body type");
     }
 
-    auto body = result.Message->GetBodyAsAmqpValue();
+    auto &body = result.Message->GetBodyAsAmqpValue();
     if (body.GetType() != Azure::Core::Amqp::Models::AmqpValueType::Map)
     {
       throw std::runtime_error("Unexpected body type");
