+++ sdk/eventhubs/azure-messaging-eventhubs/samples/produce-events/produce_events.cpp
@@ -51,27 +51,43 @@ int main()
     Azure::Messaging::EventHubs::Models::EventData event;
     event.Body = {1, 3, 5, 7};
     event.MessageId = "test-message-id";
-    batch.TryAddMessage(event);
+    if (!batch.TryAdd(event))
+    {
+      std::cerr << "Failed to add the event to the batch" << std::endl;
+      return 1;
+    }
   }
 
   {
     Azure::Messaging::EventHubs::Models::EventData event;
     event.Body = {2, 4, 6, 8, 10};
-    batch.TryAddMessage(event);
+    if (!batch.TryAdd(event))
+    {
+      std::cerr << "Failed to add the event to the batch" << std::endl;
+      return 1;
+    }
   }
 
   // Send an event with a body initialized at EventData constructor time.
   {
     Azure::Messaging::EventHubs::Models::EventData event{1, 1, 2, 3, 5, 8};
     event.MessageId = "test-message-id-fibonacci";
-    batch.TryAddMessage(event);
+    if (!batch.TryAdd(event))
+    {
+      std::cerr << "Failed to add the event to the batch" << std::endl;
+      return 1;
+    }
   }
 
   // Send an event with a UTF-8 encoded string body.
   {
     Azure::Messaging::EventHubs::Models::EventData event{"Hello Eventhubs!"};
     event.MessageId = "test-message-id-helloworld";
-    batch.TryAddMessage(event);
+    if (!batch.TryAdd(event))
+    {
+      std::cerr << "Failed to add the event to the batch" << std::endl;
+      return 1;
+    }
   }
 
   producerClient.Send(batch);
